Status for registration
alot of logic added and some problems has occured that every existing
user without the status.pending is being shown but not the new user
with the status pending. so i need to change that. When i accept
the listed users without status the program crashes so i need to show
status users and accept or decline them.