The logic for accepting and denying
The logic is now done and it took some time and when i saw a solution i
went for it. The solution was to saave the lines in a list that did not
match the selected patient email and when thats found the last index
part got changed to accepted or denied.